+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Call afterNotify handlers with error if notify preconditions fail (#654)
+- Call beforeNotify handlers even if preconditions fail (#654)
 
 ## [3.2.7] - 2021-11-01
 ### Fixed